recommendation_system/requirements.txt
# Core dependencies (versions avec wheels pour Windows)
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# NLP and Embeddings
sentence-transformers>=2.2.0
spacy>=3.5.0
transformers>=4.30.0

# Vector Search
faiss-cpu>=1.7.0

# API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# CV Parsing
PyPDF2>=3.0.0
python-docx>=1.0.0
pdfplumber>=0.10.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
joblib>=1.3.0

# Streamlit (pour l'interface web)
streamlit>=1.28.0
plotly>=5.15.0

# Optional: Téléchargez le modèle spaCy après installation:
# python -m spacy download fr_core_news_lg
